refactor TransactionHeading
