+++ packages/suite/src/views/wallet/transactions/components/TransactionList/components/TransactionItem/components/TransactionHeading/index.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
+import { colors, variables } from '@trezor/components';
+import { Translation, HiddenPlaceholder, Sign } from '@suite-components';
+import { isTxUnknown, getTargetAmount, getTxOperation } from '@wallet-utils/transactionUtils';
 import { WalletAccountTransaction } from '@wallet-types';
-import { Translation } from '@suite-components';
-import { isTxUnknown } from '@wallet-utils/transactionUtils';
 
 const Wrapper = styled.span`
     flex: 1 1 auto;
@@ -10,15 +11,58 @@ const Wrapper = styled.span`
     overflow: hidden;
 `;
 
+const CryptoAmount = styled.span`
+    color: ${colors.NEUE_TYPE_DARK_GREY};
+    font-size: ${variables.FONT_SIZE.NORMAL};
+    font-weight: ${variables.FONT_WEIGHT.MEDIUM};
+    /* line-height: 1.5; */
+    text-transform: uppercase;
+    white-space: nowrap;
+    flex: 0;
+`;
+
+const StyledHiddenPlaceholder = styled(HiddenPlaceholder)`
+    /* padding: 8px 0px; row padding */
+    display: block;
+    overflow: hidden;
+    white-space: nowrap;
+    text-overflow: ellipsis;
+`;
+
 interface Props {
     transaction: WalletAccountTransaction;
     isPending: boolean;
+    useSingleRowLayout: boolean;
 }
 
-const TransactionHeading = ({ transaction, isPending }: Props) => {
+const TransactionHeading = ({ transaction, isPending, useSingleRowLayout }: Props) => {
+    let amount = null;
+
+    if (useSingleRowLayout) {
+        const isTokenTransaction = transaction.tokens.length > 0;
+        const target = transaction.targets[0];
+        const transfer = transaction.tokens[0];
+        const targetAmount = !isTokenTransaction
+            ? getTargetAmount(target, transaction)
+            : transfer.amount;
+        const symbol = !isTokenTransaction ? transaction.symbol : transfer.symbol;
+        const operation = getTxOperation(transaction);
+        amount = (
+            <CryptoAmount>
+                {targetAmount && (
+                    <StyledHiddenPlaceholder>
+                        {operation && <Sign value={operation} />}
+                        {targetAmount} {symbol}
+                    </StyledHiddenPlaceholder>
+                )}
+            </CryptoAmount>
+        );
+    }
+
     // TODO: intl once the structure and all combinations are decided
     let heading = null;
     const symbol = transaction.symbol.toUpperCase();
+
     if (isTxUnknown(transaction)) {
         heading = <Translation id="TR_UNKNOWN_TRANSACTION" />;
     } else if (transaction.type === 'sent') {
@@ -30,7 +74,13 @@ const TransactionHeading = ({ transaction, isPending }: Props) => {
     } else {
         heading = `Unknown ${symbol} transaction`;
     }
-    return <Wrapper>{heading}</Wrapper>;
+
+    return (
+        <>
+            <Wrapper>{heading}</Wrapper>
+            {amount}
+        </>
+    );
 };
 
 export default TransactionHeading;
